extract getSalesFromCardname
